+++ Version_history.md
@@ -1,4 +1,32 @@
 
+# Version 0.0.4
+### (database handling)
+
+
+* **???** 
+
+
+
+---
+
+# Version 0.0.3
+### (start with logic for settings window)
+
+
+* **gui_manager** added all windows settings (open, close, reject)
+* **version.html** updated design (center text)
+* **gui_data**  updated some name definition in all windows
+* **settings_manager** created for settings_window logic
+* **settings.json** for file path storage
+* **settings_manager** logic for database creation (db predefined with colums and tables);
+* **settings_manager** logic for dxcc list import (always overrides existing file!!);      //FIXME: only overwrite when newer or different or hold last file as backup
+* **settings_manager** logic for download folder selection added
+* **gui_manager** function of button reset; new db; select db and text field for db path added;
+* **gui_manager** function of button select dxcc and text field for dxcc path added
+* **gui_manager** function of button select download folder and text field for download folder path added
+
+
+---
 
 # Version 0.0.2
 ### (start of coding, in different branches)